04_streamlit_app/requirements.txt
streamlit>=1.31.0
snowflake-connector-python>=3.6.0
pandas>=2.0.0
plotly>=5.18.0
numpy>=1.24.0
